requirements.txt
asgiref==3.8.1
attrs==24.3.0
autopep8==2.3.1
Django==5.1.4
django-filter==24.3
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
Faker==33.3.0
gprof2dot==2024.6.6
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.1.0
pycodestyle==2.12.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
